requirements.txt
fastapi
uvicorn[standard]
pymilvus
openai
python-dotenv
aiofiles 